dataset doc update
merge lingrui s work
